ChatBotBackkend/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.10.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pyjwt==2.9.0
ollama==0.3.3
